+++ LaunchServer/source/response/profile/BatchProfileByUsernameResponse.java
@@ -11,7 +11,7 @@ import launchserver.LaunchServer;
 import launchserver.response.Response;
 
 public final class BatchProfileByUsernameResponse extends Response {
-	public BatchProfileByUsernameResponse(LaunchServer server, int id, HInput input, HOutput output) {
+	public BatchProfileByUsernameResponse(LaunchServer server, long id, HInput input, HOutput output) {
 		super(server, id, input, output);
 	}
