+++ LaunchServer/source/response/update/UpdateListResponse.java
@@ -11,7 +11,7 @@ import launchserver.LaunchServer;
 import launchserver.response.Response;
 
 public final class UpdateListResponse extends Response {
-	public UpdateListResponse(LaunchServer server, int id, HInput input, HOutput output) {
+	public UpdateListResponse(LaunchServer server, long id, HInput input, HOutput output) {
 		super(server, id, input, output);
 	}
